// workspaces/frontend/src/generated/OpenApiTypes.ts
export interface WorkspacesActivity {
  /** Unix Epoch time */
  lastActivity: number;
  lastProbe?: WorkspacesLastProbeInfo;
  /** Unix Epoch time */
  lastUpdate: number;
}

export interface WorkspacesHttpService {
  displayName: string;
  httpPath: string;
}

export interface WorkspacesImageConfig {
  current: WorkspacesOptionInfo;
  desired?: WorkspacesOptionInfo;
  redirectChain?: WorkspacesRedirectStep[];
}

export interface WorkspacesImageRef {
  url: string;
}

export interface WorkspacesLastProbeInfo {
  /** Unix Epoch time in milliseconds */
  endTimeMs: number;
  message: string;
  result: WorkspacesProbeResult;
  /** Unix Epoch time in milliseconds */
  startTimeMs: number;
}

export interface WorkspacesOptionInfo {
  description: string;
  displayName: string;
  id: string;
  labels: WorkspacesOptionLabel[];
}

export interface WorkspacesOptionLabel {
  key: string;
  value: string;
}

export interface WorkspacesPodConfig {
  current: WorkspacesOptionInfo;
  desired?: WorkspacesOptionInfo;
  redirectChain?: WorkspacesRedirectStep[];
}

export interface WorkspacesPodMetadata {
  annotations: Record<string, string>;
  labels: Record<string, string>;
}

export interface WorkspacesPodMetadataMutate {
  annotations: Record<string, string>;
  labels: Record<string, string>;
}

export interface WorkspacesPodSecretInfo {
  defaultMode?: number;
  mountPath: string;
  secretName: string;
}

export interface WorkspacesPodSecretMount {
  defaultMode?: number;
  mountPath: string;
  secretName: string;
}

export interface WorkspacesPodTemplate {
  options: WorkspacesPodTemplateOptions;
  podMetadata: WorkspacesPodMetadata;
  volumes: WorkspacesPodVolumes;
}

export interface WorkspacesPodTemplateMutate {
  options: WorkspacesPodTemplateOptionsMutate;
  podMetadata: WorkspacesPodMetadataMutate;
  volumes: WorkspacesPodVolumesMutate;
}

export interface WorkspacesPodTemplateOptions {
  imageConfig: WorkspacesImageConfig;
  podConfig: WorkspacesPodConfig;
}

export interface WorkspacesPodTemplateOptionsMutate {
  imageConfig: string;
  podConfig: string;
}

export interface WorkspacesPodVolumeInfo {
  mountPath: string;
  pvcName: string;
  readOnly: boolean;
}

export interface WorkspacesPodVolumeMount {
  mountPath: string;
  pvcName: string;
  readOnly?: boolean;
}

export interface WorkspacesPodVolumes {
  data: WorkspacesPodVolumeInfo[];
  home?: WorkspacesPodVolumeInfo;
  secrets?: WorkspacesPodSecretInfo[];
}
// ...
export interface WorkspacesService {
  httpService?: WorkspacesHttpService;
}

export interface WorkspacesWorkspace {
  activity: WorkspacesActivity;
  deferUpdates: boolean;
  name: string;
  namespace: string;
  paused: boolean;
  pausedTime: number;
  pendingRestart: boolean;
  podTemplate: WorkspacesPodTemplate;
  services: WorkspacesService[];
  state: WorkspacesWorkspaceState;
  stateMessage: string;
  workspaceKind: WorkspacesWorkspaceKindInfo;
}

export interface WorkspacesWorkspaceCreate {
  deferUpdates: boolean;
  kind: string;
  name: string;
  paused: boolean;
  podTemplate: WorkspacesPodTemplateMutate;
}

export interface WorkspacesWorkspaceKindInfo {
  icon: WorkspacesImageRef;
  logo: WorkspacesImageRef;
  missing: boolean;
  name: string;
}

export enum WorkspacesWorkspaceState {
  WorkspaceStateRunning = 'Running',
  WorkspaceStateTerminating = 'Terminating',
  WorkspaceStatePaused = 'Paused',
  WorkspaceStatePending = 'Pending',
  WorkspaceStateError = 'Error',
  WorkspaceStateUnknown = 'Unknown',
}